+++ spec/data_porter/dsl/column_spec.rb
@@ -1,6 +1,20 @@
 # frozen_string_literal: true
 
 RSpec.describe DataPorter::DSL::Column do
+  describe "synonyms" do
+    it "defaults to empty array" do
+      column = described_class.new(name: :email)
+
+      expect(column.synonyms).to eq([])
+    end
+
+    it "accepts an array of synonyms" do
+      column = described_class.new(name: :email, synonyms: %w[correo adresse_email])
+
+      expect(column.synonyms).to eq(%w[correo adresse_email])
+    end
+  end
+
   describe "transform" do
     it "defaults to empty array" do
       column = described_class.new(name: :email)